change datatypes in sensordata model and add parser for incoming data

+++ qt/application/QT_GUI/Diagram/customplotbarchart.cpp
@@ -106,7 +106,7 @@ void CustomPlotBarChart::calculateData()
         const SensorData* data = model.getSingleSensorData(index);
         if(data != 0)
         {
-            m_yAxis.append(data->getHeartRate().toInt());
+            m_yAxis.append(data->getHeartRate());
             time = time + 1;
         }
         else
